Player Movement
Updated the ability of players to move based on a random roll of a die from 1 to 3. Made the random number 1 to 3 as the size of the board is small. Haven't implemented edge cases, so if the player tries to move to a location that doesn't exist, an error occurs.